refactor the common media fetching logic out
